Moved the code for spectroscopic sight lines to observers/spectroscopy and for group observers to observers/group. Added documentation.
