+++ cherab/tools/observers/group/__init__.py
@@ -1,7 +1,7 @@
 
-# Copyright 2016-2018 Euratom
-# Copyright 2016-2018 United Kingdom Atomic Energy Authority
-# Copyright 2016-2018 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+# Copyright 2016-2021 Euratom
+# Copyright 2016-2021 United Kingdom Atomic Energy Authority
+# Copyright 2016-2021 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
 #
 # Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
 # European Commission - subsequent versions of the EUPL (the "Licence");
@@ -17,8 +17,5 @@
 # See the Licence for the specific language governing permissions and limitations
 # under the Licence.
 
-from .bolometry import BolometerCamera, BolometerFoil, BolometerSlit
-from .calcam import load_calcam_calibration
-from .intersections import find_wall_intersection
-from .spectroscopy import SpectroscopicSightLine, SpectroscopicFibreOptic
-from .group import SightLineGroup, FibreOpticGroup
+from .fibreoptic import FibreOpticGroup
+from .sightline import SightLineGroup
