+++ docs/source/tools/observers.rst
@@ -61,3 +61,40 @@ bolometer etendue :math:`G`, which is given by:
 
 .. autoclass:: cherab.tools.observers.bolometry.BolometerFoil
    :members:
+
+.. _observers_spectroscopic:
+
+Spectroscopic lines of sight
+----------------------------
+
+Spectroscopic line of sight allows to control main parameters of the pipeline
+without accessing the pipeline directly. Multiple spectroscopic line of sight can be
+combined into a group.    
+
+.. autoclass:: cherab.tools.observers.spectroscopy.base._SpectroscopicObserver0DBase
+   :members:
+
+.. autoclass:: cherab.tools.observers.spectroscopy.SpectroscopicSightLine
+   :members:
+
+.. autoclass:: cherab.tools.observers.spectroscopy.SpectroscopicFibreOptic
+   :members:
+
+.. _observers_group:
+
+Group observers
+---------------
+
+Group observer is a collection of observers of the same type, for example an array of line of
+sight. The parameters of individual observers in a group may differ.
+Group observer allows combined observation, namely, calling the observe function for a group
+leads to a sequential call of this function for each observer in the group.
+
+.. autoclass:: cherab.tools.observers.group.base.Observer0DGroup
+   :members:
+
+.. autoclass:: cherab.tools.observers.group.SightLineGroup
+   :members:
+
+.. autoclass:: cherab.tools.observers.group.FibreOpticGroup
+   :members:
